create_form/refactor form
